Integrate Room database and improve data handling
- Adds Room dependencies and configures kapt for annotation processing.
- Adds Room database configuration and ElementDao for data access.
- Creates OverpassQueryBuilder for dynamic query building based on categories.
- Updates the Element and Tags objects.
- Creates Converter class to handle JSON serialization for Tags objects.
- Updates PlaceRepository to cache data locally using Room.
- Creates own BoundingBox class and refactors MapViewModel to use this bounding box and load markers efficiently.
- Updates DI module to include Room database and DAO setup.

Room in this case is used for local data chaching to store markers beyond the current zoom level.
This should reduce the need for frequent API calls. However, the zoom levels and all are still horribly adjusted in this commit.
Instead, it might be more efficient to use Overpass API selectively to load detailed information only when a marker is clicked.