Refactor and enhance map functionality
- Refactors zoom/move intents into a single Move intent.
- Updates Room database to work in-memory.
- Removes custom BoundingBox model and uses osmdroid BoundingBox for easier calculations.
- Renames models from Element to Marker.
- Adds OkHttpClient in AppModule for network operations.
- Adds Bbox parameter for Move Intent.
- Creates OverpassApiResponse model for handling API responses.
- Adds new states to MapState to handle logic for snapshots of the latest Bounding Box.
- Adds logging for easier debugging.
- Creates MapListener to handle custom onZoom and onScroll events.
- Refactors MapViewModel logic.

API calls and cached markers will be fetched only when the center is out of the latest bounding box (which is the latest snapshot of the screen and its area). To have some buffer, EXPAND_FACTOR is used to fetch markers even beyond the current screen. Currently, markers load but zooming out is not handled at all. Additionally, the expand factor and zoom threshold should be tested for more optimal values. Marker fetching is still slow when moving the map quickly, and the expand factor seems to be too small.

Next steps would be to add tests, handle zooming, and remove the suppress tags.