Create more customized data class for places
Creates a Place data class with more specific and customized data fields to better suit the purpose of the app. This includes more specific accessibility information. The Place class will serve as the main data class of the project so it can be used across various views instead of MapMarker, which name is not as versatile either.