Add custom marker info window with preliminary accessibility info
Adds new custom info popup window when a cluster item is clicked. Currently, there is an issue regarding cluster item click which cannot take Marker composable but instead just basic composables. Current solution was to insert own custom info window just above the currently selected cluster item. This should be changed however in the future when Google fixes this issue regarding onClusterItemClick.

Adds new string resources, and updates move handling and API calls to use separate flow scopes, since move intent should not be rebounced but instead the API calls should be. This was needed because of the custom made and inserted info window which would not stay on the correct spot if debounce was used for move intents as well.

Part of #3, #10