Setup initial connection to Supabase
Adds functionality to fetch places from Supabase and display on the map.
Includes data models for places and API service to interact with Supabase.

This commit gets the Supabase connection working and fetches data successfully. Refining the logic will come later as part of this larger refactor. Lot of earlier code is simplified and left out.

Part of #40