Refactor map data loading and filtering
Major refactor to the logic for loading and managing places on the map. Currently, captures a snapshot of the larger area when zoom level is over the threshold , fetches all places from the API for those bounds, inserts them into Room, and streams data from Room based on the current map bounds to the ViewModel, which updates the state accordingly. The ViewModel applies a limit on the maximum number of places displayed in the UI, and the state keeps track of "cached" places to filter out duplicates when fetching new data from the API and inserting it into Room.

However, the logic to avoid fetching data for areas that have already been retrieved is incomplete. While the insertion logic correctly prevents adding duplicates to Room, the API may still fetch the same areas multiple times due to missing filtering logic for already-fetched places.

Additionally, FTS is currently disabled and needs to be reintroduced later. The primary issue with re-enabling FTS is the requirement for an integer rowId in the FTS table, while our domain model uses a String for the id, matching the database schema (the uuid). This conflict needs to be resolved before FTS can be fully functional again.

Part of #40