Add accessibility status display to details view
Adds an accessibility status display section to the place details bottom sheet. It displays a status icon or a view for the user to choose the accessibility of the place if it is unknown. Also, add icons and strings to support the new feature. Adds handling the sheet peek height dynamically.

Part of #49