Fix issue with updating entries
Fixes issue where entries were not updated to db from client side. For now uses two separate API calls to update the property and user_modified flag as there seems to be an issue with serialization preventing the mix of different value types (Boolean and String in this case) in a single mapping.